pyproject.toml
[project]
name = "megano-online-store-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "uservm",email = "example@mail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "django-filter (>=25.1,<26.0)",
    "diploma-frontend @ file:///home/uservm/PycharmProjects/megano_online_store_2025/diploma-frontend/dist/diploma-frontend-0.6.tar.gz",
    "pillow (>=11.2.1,<12.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
